Fix the endless map creation without running a map thing.
Klusark's fix was not the full story. This block of code was expecting
Either A) shouldfarm is true, B) worldlevel == highestmaplevel otherwise
it could C) Create.   Therefore leading to the endless loop of creation.
As tested with this savegamefile: http://pastebin.com/yrRLhb8y  and the
Prestige setting set to "OFF" (this easiest way to replicate a a "Not
enough health|Damage" scenario)

This entire system of map creation in 1 cycle  and running it in another
cycle is really hard to read the code of. If Siphonmap ==-1, it changes
shouldDoMap to "Create" and skips the block right after, and goes on to
create the map. Then in the next cycle, siphonmap should be something
not -1, and this code block will run, and set ShouldDoMap to the map we
created last cycle (which it has to re-find all over again) using either
[highestmap] or [siphonmap] code. (it finds this about 100 lines up).